fix-webapp: fix dateTimeRangePicker by removing seconds from minDateTime props
